+++ SFZ_MIDI_Player/SFZ_MIDI_Player.vcxproj
@@ -116,12 +116,13 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="Main.cpp" />
-    <ClCompile Include="source\AudioLoader.cpp" />
+    <ClCompile Include="source\AudioLoadManager.cpp" />
     <ClCompile Include="source\MIDILoader.cpp" />
     <ClCompile Include="source\PianoRoll.cpp" />
     <ClCompile Include="source\SamplePlayer.cpp" />
     <ClCompile Include="source\SampleSource.cpp" />
     <ClCompile Include="source\SFZLoader.cpp" />
+    <ClCompile Include="source\WaveLoader.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
@@ -322,13 +323,15 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\Animation.hpp" />
-    <ClInclude Include="include\AudioLoader.hpp" />
+    <ClInclude Include="include\AudioLoaderBase.hpp" />
+    <ClInclude Include="include\AudioLoadManager.hpp" />
     <ClInclude Include="include\MIDILoader.hpp" />
     <ClInclude Include="include\PianoRoll.hpp" />
     <ClInclude Include="include\SamplePlayer.hpp" />
     <ClInclude Include="include\SampleSource.hpp" />
     <ClInclude Include="include\SFZLoader.hpp" />
     <ClInclude Include="include\Utility.hpp" />
+    <ClInclude Include="include\WaveLoader.hpp" />
     <ClInclude Include="stdafx.h" />
   </ItemGroup>
   <ItemGroup>
