+++ SFZ_MIDI_Player/source/SamplePlayer.cpp
@@ -5,7 +5,7 @@
 #include <PianoRoll.hpp>
 #include <MIDILoader.hpp>
 #include <SampleSource.hpp>
-#include <AudioLoader.hpp>
+#include <AudioLoadManager.hpp>
 
 namespace
 {
